Renamed watchlist middleware
