+++ src/modules/pos/validator/store/middlewares/validators.js
